[Refactor] Player 클래스 생성 및 로직 이동
