+++ src/main/kotlin/lotto/Application.kt
@@ -1,20 +1,9 @@
 package lotto
 
-import camp.nextstep.edu.missionutils.Console
-
 fun main() {
-    print(inputPurchaseAmount())
+    val player = Player()
+    player.inputPurchaseAmount()
 }
 
-fun inputPurchaseAmount(): Int {
-    println("구입금액을 입력해 주세요.")
-    val amount = Console.readLine().toInt()
-    validatePurchaseAmount(amount)
-    return amount
-}
-
-fun validatePurchaseAmount(amount: Int) {
-    require(amount % 1000 == 0) { throw IllegalArgumentException("[ERROR] 1,000원 단위로 입력해주세요.") }
-}
